Reverse bit order
Other systems (such as Linux) have reversed bit order for words from what out
code is doing. Change to the convention that the high bit in a word is the first
bit. This actually removes reversing code from much of the code and adds
a reverse in the chien registers.

This makes the output match the Linux software BCH codes.